Rename component to PageCTA
