+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 // src/App.js
 import HomeHero from 'organisms/HomeHero';
-import ContactCTA from '../components/organisms/ContactCTA';
+import ContactCTA from '../components/organisms/PageCTA';
 import ServicesSection from 'organisms/ServicesSection';
 import WhySection from 'organisms/WhySection';
