+++ src/pages/Services.jsx
@@ -1,5 +1,5 @@
 import ServicesHero from '../components/organisms/ServicesHero';
-import ContactCTA from '../components/organisms/ContactCTA';
+import ContactCTA from '../components/organisms/PageCTA';
 import ServicesInfo from '../components/organisms/ServicesInfo';
 
 const Services = () => {
